packaging: refactor tests into a single directory
