+++ cmd/grafana-agent/run-config.river
@@ -1,0 +1,5 @@
+server {
+	http {
+		listen_addr = "127.0.0.1:12381"
+	}
+}
